final_pjt_back/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.5
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.14.0
django-extensions==3.2.0
django-seed==0.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.2
executing==1.0.0
Faker==13.3.4
idna==3.4
inflection==0.5.1
ipython==8.4.0
jedi==0.18.1
jsonschema==4.17.3
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.0
toposort==1.7
traitlets==5.3.0
tzdata==2023.2
uritemplate==4.1.1
urllib3==2.0.2
wcwidth==0.2.5
